refac: Move customer named files to account named files.
